Moved CLI demo to CommandLine dir.
