+++ demos/CommandLine/CommandLine.csproj
@@ -20,6 +20,6 @@
 
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\PowerSync\Powersync.Common\PowerSync.Common.csproj" />
+    <ProjectReference Include="..\..\PowerSync\Powersync.Common\PowerSync.Common.csproj" />
   </ItemGroup>
 </Project>
