+++ demos/CommandLine/NodeConnector.cs
@@ -1,4 +1,4 @@
-﻿namespace CLI;
+﻿namespace CommandLine;
 
 using System;
 using System.Collections.Generic;
