+++ root.sln
@@ -7,17 +7,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "PowerSync", "PowerSync", "{
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PowerSync.Common", "PowerSync\PowerSync.Common\PowerSync.Common.csproj", "{EA0A66FF-9E8A-4465-84EC-B2CFFC7E4393}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "demos", "demos", "{6A2C8920-FA72-44E3-AB08-4FC8B2655A30}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Command-Line", "Command-Line", "{461A9BB1-D773-451F-BD3E-4765A332A782}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CLI", "demos\Command-Line\CLI\CLI.csproj", "{42BC9844-F934-4C26-9354-0B78BF770DD2}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{7404B997-119F-4826-845A-18023CC9F002}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "PowerSync", "PowerSync", "{C784FBE4-CC1E-4A0A-AE8E-6B818DD3724D}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PowerSync.Common.Tests", "Tests\PowerSync\PowerSync.Common.Tests\PowerSync.Common.Tests.csproj", "{11102D8A-1B2C-4299-8289-88909F132A56}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "demos", "demos", "{9144195A-C68F-4B1E-A574-474EDD424D6C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommandLine", "demos\CommandLine\CommandLine.csproj", "{D7FDA714-D29F-4D85-B3F2-74C6810A36F9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -32,20 +30,19 @@ Global
 		{EA0A66FF-9E8A-4465-84EC-B2CFFC7E4393}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EA0A66FF-9E8A-4465-84EC-B2CFFC7E4393}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EA0A66FF-9E8A-4465-84EC-B2CFFC7E4393}.Release|Any CPU.Build.0 = Release|Any CPU
-		{42BC9844-F934-4C26-9354-0B78BF770DD2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{42BC9844-F934-4C26-9354-0B78BF770DD2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{42BC9844-F934-4C26-9354-0B78BF770DD2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{42BC9844-F934-4C26-9354-0B78BF770DD2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{11102D8A-1B2C-4299-8289-88909F132A56}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{11102D8A-1B2C-4299-8289-88909F132A56}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{11102D8A-1B2C-4299-8289-88909F132A56}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{11102D8A-1B2C-4299-8289-88909F132A56}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D7FDA714-D29F-4D85-B3F2-74C6810A36F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D7FDA714-D29F-4D85-B3F2-74C6810A36F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D7FDA714-D29F-4D85-B3F2-74C6810A36F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D7FDA714-D29F-4D85-B3F2-74C6810A36F9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{EA0A66FF-9E8A-4465-84EC-B2CFFC7E4393} = {B1D87BA9-8812-4EFA-BBBE-1FF1EEEB5433}
-		{461A9BB1-D773-451F-BD3E-4765A332A782} = {6A2C8920-FA72-44E3-AB08-4FC8B2655A30}
-		{42BC9844-F934-4C26-9354-0B78BF770DD2} = {461A9BB1-D773-451F-BD3E-4765A332A782}
 		{C784FBE4-CC1E-4A0A-AE8E-6B818DD3724D} = {7404B997-119F-4826-845A-18023CC9F002}
 		{11102D8A-1B2C-4299-8289-88909F132A56} = {C784FBE4-CC1E-4A0A-AE8E-6B818DD3724D}
+		{D7FDA714-D29F-4D85-B3F2-74C6810A36F9} = {9144195A-C68F-4B1E-A574-474EDD424D6C}
 	EndGlobalSection
 EndGlobal
